+++ src/components/BlurImage/BlurImage.tsx
@@ -1,20 +1,9 @@
 import Image from 'next/image'
 import { useState } from 'react'
 
-import { css } from '~/styles'
-
 import type { ImageProps } from 'next/image'
 
-const styles = {
-  img: css({
-    transitionDuration: '0.5s',
-    transitionTimingFunction: 'ease-in-out'
-  })(),
-  imgBlur: css({
-    willChange: 'transform',
-    transform: 'scale(1.05)'
-  })()
-}
+import styles from './BlurImage.css'
 
 interface Props {
   blurClassName?: string
