+++ src/components/blog/index.ts
@@ -1,2 +1,2 @@
-export { default as Hero } from './Hero'
-export { default as PostCard } from './PostCard'
+export { default as Hero } from './Hero/Hero'
+export { default as PostCard } from './PostCard/PostCard'
