+++ src/components/index.ts
@@ -1,7 +1,7 @@
 export { default as Header } from './Header/Header'
 export { default as Footer } from './Footer'
 export { default as Article } from './Article'
-export { default as BlurImage } from './BlurImage'
+export { default as BlurImage } from './BlurImage/BlurImage'
 export { default as VisuallyHidden } from './VisuallyHidden'
 export { default as ThemeToggle } from './ThemeToggle/ThemeToggle'
 export { default as ScrollArea } from './ScrollArea'
